moved indexedDbStorage to separate file
