+++ redisinsight/ui/src/services/index.ts
@@ -10,3 +10,4 @@ export * from './theme'
 export * from './hooks'
 export * from './capability'
 export { apiService, resourcesService }
+export { IndexedDbStorage } from 'uiSrc/services/indexedDbStorage';
